Use remote instead of local salt API, separate modules.
This means that the script does not have to be run on the salt master,
but can communicate to one over HTTPS. The master has to be configured
to receive external commands using cherrypy.

For this, a thin wrapper was written around the cherrypy salt API that
only exposes commands needed by bootstrap.

Additionally, the names of arguments were changed to be more generic, in
an effort to create kind of a plugin API for the installation server,
the hypervisor and the configuration management system, so one can later add
support for other systems more easily.

For the hypervisor, such a generic API was created and an implemented
for libvirt.